move rendered content styles to RenderContent component
