+++ components/Post/index.tsx
@@ -1,4 +1,5 @@
 /* eslint-disable @next/next/no-img-element */
+import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import dayjs from 'dayjs'
@@ -13,17 +14,14 @@ import { Subscribe } from '@components/Subscribe'
 import { TableOfContents } from '@components/toc/TableOfContents'
 
 import { SEO } from '@meta/seo'
+import { ISeoImage } from '@meta/seoImage'
 
 import { PostClass } from '@helpers/PostClass'
 import { GhostPostOrPage, GhostPostsOrPages, GhostSettings } from '@lib/ghost'
 import { collections } from '@lib/collections'
 
-import { ISeoImage } from '@meta/seoImage'
-
-import React from 'react'
 import { Header } from '../Header'
 import { Heading1 } from '../typography/Headings'
-import { PostFullContent } from './components'
 
 interface PostProps {
   cmsData: {
@@ -107,9 +105,7 @@ export const Post = ({ cmsData }: PostProps) => {
                   </div>
                 </header>
 
-                <PostFullContent>
-                  <RenderContent htmlAst={htmlAst} />
-                </PostFullContent>
+                <RenderContent htmlAst={htmlAst} />
 
                 <div className="post-meta">
                   {featImg &&
