+++ components/RenderContent/components/iframe.tsx
@@ -1,0 +1,6 @@
+/* Override third party iframe styles */
+export const iframeStyles = `
+  iframe {
+    margin: 0 auto !important;
+  }
+`
