+++ components/RenderContent/components/image.tsx
@@ -25,11 +25,3 @@ export const imageCaptionStyles = `
     text-align: center;
   }
 `;
-
-/* Override third party iframe styles */
-export const iframeStyles = `
-  iframe {
-    margin: 0 auto !important;
-  }
-`
-
